fix: rollback package separation get-webpack-alias
